+++ frontend/src/services/api.js
@@ -2,9 +2,9 @@
 // Por eso es "async", necesitamos decirle a JS "espera a que termine"
 export async function generateContent(datos) {
 
-    const apiBaseUrl = process.env.REACT_APP_API_URL || 'http://localhost:5000/api';
+    const apiBaseUrl = process.env.REACT_APP_API_URL || 'http://localhost:5001/api';
     const url = `${apiBaseUrl}/generate`;
-    
+
     try { // maneja errores
         const response = await fetch(url, { // hace la peticion y response lo que vuelve del servidor
             // await espera (porque es async)